feat(Game): Making Chess component responsable
The Chessboard now resizes to the windowwith if the windowwith is below 700px above that the width stays the same.
The windowwith now gets stored inside the Store to make it more accesable. The Maincomponent listens for resizes and updates the store if event accures.
Starting to convert the chessboard styles to scss.